pyproject.toml
[tool.poetry]
name = "drf-owu"
version = "0.1.0"
description = ""
authors = ["vladzo <vashchukvlad17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.4"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
